move GenesisStateWithSingleValidator func to tests/e2e/baseapp
